refactor: templates로 폴더 변경
